added hooks useRedirectAuth in sign in, sign up and forgot password pages
